disk: Read attach field using volume's Virtualmachineid
The `.Attached` field is unreliable, there are situations where
cloudstack will return an empty `.Attached` value even if the disk is
attached to a VM. So far I've seem this happen when the disk is created
alongside with the virtualmachine (deployVirtualMachine called with the
`size` attribute) but there may be other situations where this also
happens.

Checking for the `.Virtualmachineid` field to set the `attach` attribute
seems a safer approach as it's always set when the volume is already
attached.

I came across this when importing an existing infrastructure to
terraform by manually creating a state file, the existing VMs have
attached volumes that were created with them and these volumes don't
have a `.Attached` field. This causes the cloudstack provider to try to
attach them to the same VMs they are already attached, raising an error.